requirements.txt
asgiref==3.11.0
beautifulsoup4>=4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-environ==0.12.0
djangorestframework==3.16.1
docker==7.1.0
idna==3.11
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.2
markdown>=3.10
Pillow>=10.0.0
